pyproject.toml
[build-system]
requires = ["setuptools>=75", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fek-extractor"
version = "0.1.0"
description = "Extract structured data from FEK PDFs."
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Your Name", email = "you@example.com" }]
dependencies = [
  "pdfminer.six>=20221105"
]
[project.optional-dependencies]
dev = [
  "pytest>=8",
  "pytest-cov>=5",
  "ruff>=0.5",
  "black>=24.8",
  "mypy>=1.10",
  "pre-commit>=3.7",
  "mkdocs>=1.5",
  "mkdocs-material>=9.5",
  "build>=1.2.1",
]
[project.scripts]
fek-extractor = "fek_extractor.cli:main"

[tool.setuptools]
package-dir = { "" = "src" }
[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"
[tool.ruff.lint]
select = ["E","F","I","UP","B","SIM"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.mypy]
python_version = "3.10"
strict = true
mypy_path = ["src"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
